Utilize unified prob threshold in factoid skill.
Before these changes, 2 different thresholds for factoid were set. One threshold 0.96 to turn the skill on, another 0.9 in the same skill. Given that the current model classifies question "where is montreal" as factoid with 0.55 probability, I have set the threshold 0.5 for both of these cases. And made the one threshold to use in both places to streamline the code